Adding an option to run test with multiple threads sharing a worker.
Move down the thread creation logic so as not to go MT on rte

Allocate enough buffers for all threads to have something to chew on

Abide to the coding style

minor: changin the name of field thread_num to thread_count

Missing thread_count=1 in the other test

Removing global variables and other comments from Yosefe.

Using the right license snippet and adding to the LICENSE file.

indentation.

Fix a leak found by coverity