UCT/BASE/SM: Refactor mm to be a transport.
Implement mm as a transport which can use either "mmap" or "sysv" pd.
 Instead of using mm from sysv, create a small framework for "mm
components", which have a basic alloc/attach/release functionality.
sysv is an "mm component", and in the future so will be mmap. mm
implements the complete set of tl and pd operations on top of the mm
component interface. sysv and mmap will still be separate pd components
because the have different ways of doing rkey_unpack.
 Some changes in the infrastreucture were required: Adding "priv" filed
to pd component (to hold mm mapper operations), and allowing single
transport be used from different pd components.