UCM/TEST: Add UCM library which provides memory (un)mapping hooks.
 libucm.so is a standalone non-unloadable library which installs hooks
for virtual memory changes in the current process. Then, it calls
user-defined callbacks, which may potentially override the default
behavior, or just passively listen and update their own data.
 libucm.so does not use libuct.so, to avoid making it non-unloadable as
well, and impelements a basic logging service which is safe to use from
malloc hooks.

The hooks employed so far are (in order of initialization):
  1. override mmap/munmap/mremap/... symbols.
  2. modify relocation tables of loaded objects (got) to point to
     out mmap/munmap/... functions.
  3. install malloc hooks which use our mmap/munmap/...
Only the minimal set of hooks which satisfies all required events are
installed.