UCP/UCT/API/TEST: Allow pending callback to return UCS_INPROGRESS.
 In order to achieve better fairness between progressing requests on
different endpoints, and also between sends and receives, the pending
request callback should do a portion of work every time, and not
exshaust all available resources. So we add another possible return
value to the callback - UCS_INPROGRESS - which means some work was done,
but not completed.

The changes are:
- UCT API change - document new semantic
- Transports to handle the new return code
- UCP wireup to handle this case when dispatching requests from
  dummy_ep.
- Add unit test.