uct/ugni: always call GNI_CqGetevent on TX CQ
Its not safe to use external book keeping to
conditionally call GNI_CqGetEvent on TX CQs.

Particularly if GNI SMSG functions
are returning GNI_RC_NOT_DONE, that may imply
a depletion of internal CQ entry related resourcs
inside GNI.  This can prevent things like return
of back credits in one way SMSG traffic.

This commit unblocks one of the send-on-way
am tests when using GNI uct.  The test subsequently
fails with some kind of object-not-returned-to-mpool
problem, but that doesn't appear to be related to
the gni transport.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>