UCT/UD: fixes problem with simulatenous connection requests
If an endpoint has sent CREQ treat incoming CREQ as a CREP.
Add assertion that checks that CREQ is never resent on already
connected endpoint.

In debug mode pass a peer name/pid during connection establishment
and keep it as part of ep. Also keep track of detailed ep state.