TOOLS/PERF: Fix address exhange when size is not same on all endpoints.
 Address and rkey size may differe between endpoints. So we have to
handle this case in perf testing tool. We change the RTE interface to
receive a single buffer instead of an iovec, since now we can't tell the
exact size of each iovec element. We also add a small header to the RTE
data, which contains the sizes of address and rkey, so the receiver
could "unpack" it correctly.