XPMEM: Fixing segfault in the disconnect flow
XPMEM unregister actually invalidates the memory for remote
access. Once the memory unregistered we can't touch
the control structure and therefor it is unsafe for caching
socket info. This patch modifies the code so the control
structure is only used for passing the information around,
while the socket info is cached on ep.

Signed-off-by: Pavel Shamis (Pasha) <pasharesearch@gmail.com>