GTEST: don't use rndv in the probe test which uses a blocking send.
the probe test uses a blocking send which cannot work in gtest with rndv
since the recv side needs to be progressed to make the send_req become
completed.
increase the rndv threshold for this test so that rndv wont be used
here. (rndv has its own gtest for testing probe)