PERF: Fixing unaligned access bug
This is a bug that got exposed by OMPI testing on ARMv8 platform.
Offset is a local variable that has never been initialized. Since it is
used for a remote address access it may lead to various unpleasant bugs.
The bug is not ARMv8 specific but for some reason it was not exposed
on other systems.

Signed-off-by: Pavel Shamis (Pasha) <pasharesearch@gmail.com>