UCP/RMA/AMO: Cache remote key translation.
In order to improve RMA and AMO performance, cache the translation of
ucp_rkey to (uct_rkey,lane) in the ucp_rkey structure. As long as the
endpoint lane configuration has not changed, the cached values remain
valid.
As part of this commit, the RMA flows are refactored and simplified.