UCP/TEST: Adaptive progress optimization.
Use wakeup mechanism to enable progress only on the interfaces which are
used for traffic. When an interface which supports wakeup is created, it
is removed from progress chain and is amed for active message events. If
an active messag arrives on that interface, it is added to the progress
chain.
In addition, an interface is added to the progress chain when it has
active endpoints.