TEST/UCP: Allow overriding endpoint and worker parameters at runtime.
In the current testing infrastructure, worker and endpoint parameters
are generated during static construction time. This does not allow
passing parameters which would only be known later, for example
user-defined context.
This change adds virtual functions which can be overridden to generate
the desired parameters. During static construction, only the "variant"
number is generated. That number can be used later to decide which
parameters to return, based to the current test variant.