UCP/TAG: Fix error handling for pending zero-copy reqeusts.
When a zero copy request is prged from the pending queue in case of
error, the offset is less than length, so the request is never
completed. Fix is to fast-forward request state when purging the pending
queue, as-if all data has been sent, and only then call the completion
handler with the error status.