TEST/FENCE: Do not connect back from receiver to sender.
Creating such connection may lead to a deadlock: the sender may get the
CONN_REQ from receiver, send a REPLY, and then wait for ACK. Since ACK
is sent on the main thread, it's possible the receiver might not send it
before the sender goes inside a blocking atomic operations. So the
sender cannot complete the wireup process and hangs.