UCT/UD: Fix race condition of arming the interface
If the progress thread processes some packets while ud_iface_event_arm
is called on the main thread, it may advance the CQ consumer index
after main thread already checked that the rx pending queue is empty.
So the progress thread will add a new pending rx, the main thread will
arm with the new consumer index, and the user will never get event a
notification for that packet.
Fix is to block the progress thread while checking pending queue and
arming the CQ.