UCP/GTEST: Force creating an am_lane for the client-server flow.
+ When using the client-server flow, and the RMA/ATOMIC features are
requested, need to make sure that ucp-wireup messages would have a lane
to be sent on. After this change, the client-server flow will require
the creation of the am_lane, both on the client and server sides.
The actual need is to have a wireup lane, but in case the wireup
lane isnt created, it will use the am_lane instead.

+ Test the above in gtest.

+ Fix a bug in the proxy ep handling. If a proxy_ep points to a
wireup_ep that is going to be destroyed (due to switching to a real
transport), make the proxy_ep point to the real transport. Don't leave
the proxy_ep pointing to a destroyed wireup_ep.

+ Add the rdmacm transport to the shared-memory transports aliases to
enable the client-server flows in shared-memory as well.