UCT/IB/MLX5: Fix TX completion accounting
Because of invalid accounting of tx completions, we could end up adding
negative number to cq_available leading to potential cq overflow, or
wraparound to negative values. When polling TX completions for mlx5
transports, we only know how many BBs were released *before* the
current WQE, but when posting we substract the number of BBs in the
*current* WQE. As a result the polling function may release more
resources than were consumed, and in the next poll "bb_num" could be
calculated to a negative value.
The solution is that the post_send function will return the number of
BBs consumed by the previous WQE instead of the current. Also we add
assertion to protect against negative values and CQ/WQ overflow.
Without the fix, the assertions would fail during some of the UCP
tests, for example send_generic_recv_contig_unexp_sync_rndv_truncated.