UCP/GTEST: Client-server - handle a case of a large worker address.
In the client-server flow, during connection establishment, the
client may have a large worker address which may not fit into the
sockaddr transport's private data length and therefore, the connection
establishment flow needs to be changed.
In this case, the client will start by sending a partial address to the
server via the sockaddr transport.
This partial address will hold the addresses of transports which are
assumed to be present in most cases (ud or ud_x or both).
Once the server side receives this message from the client, it will send
a new wireup message, which is added in this commit. This message will
hold the server's full worker address.
When the client gets this message, it will initial the regular wireup
flow by sending a wireup connection request to the server.