UCP/CONFIG: Fix support for UCTs w/o AM short
There is the transport (UCT/tcp) that doesn't have Eager short
support. This should be handled in the UCP config print operation.

The print operation doesn't work correctly, because the maximum
Eager short attribute is `ssize_t` (this is `-1`, when it is not
supported by some transport) and is passed as a `size_t` into
`ucp_ep_config_print_tag_proto`.

The fix is to change argument from `size_t` to `ssize_t` and
don't use in the comprasions with unsigned attributes when
this value is negative.