UCP/TEST: Support multiple listening transports per listener.
In client-server mode, allow for more than one listening transport on
the server.
The server will listen on all the sockaddr transports that can listen
on the given IP and port and the client would use a sockaddr transport
according to the one set by the environment.
We currently rely on the fact that the server is listening on the
transport the client chose.