UCP/WIREUP: Print better error in case wireup msg lane not found
If user has specified UNIFIED_MODE=y, but the fabric is not actullay
unified, e.g some of the machines are using emulation layer while others
don't, it's possible to get an error while trying to find an active
message lane for emulation layer wireup protocol. This lane would not be
created on a machine which did not plan to use emulation layer at all.
In this case print a better error message with a workaround suggestion.