UCP/RKEY/TEST: Fix rkey unpack and release thread safety
Move remote keys memory pool from context to worker, since the context
lock takes effect only when used within a worker lock (since it's not
activated when only one multi-threaded worker is created). Since that
memory pool has be protected by worker lock, we need to be able to
access the worker during rkey destroy, so the mpool owner must be the
worker.