UCP/FLUSH: Fix fast-forward of flush requests
When we remove a flush request from pending queue in case of error, it
can still have reference from UCT, on the lanes it waits for completion.

So we can advance the request on the lane we removed it from, and not
start it on additional lanes, but still must wait for UCT completions to
arrive.