UCS/RCACHE/TEST: Fix memory leak of rcache invalidation entries
When memory events (such as a malloc/free loop) are generated but no
rcache operations are done, the memory invalidation entries generated by
the UCM event handler are not consumed, which causes a memory leak.

To fix this issue, try to handle the memory invalidation immediately
inside the UCM event handler by removing the relevant regions from the
page table. Since we cannot completely release the region, put it on a
GC list. This will not cause memory leak because it's bounded by the
number of valid regions.

The tests also have to be adjusted, to expect memory regions to be
invalidated immediately (unless under lock).