UCT/UD/PERF: Fix uct_ep_flush() for UD
In order to make sure flush is completed after all outstanding resends
are completed, maintain strict order in the send window: release skb's
only after they are both acknowledged and not being resent anymore. It
simplifies the flush and release logic, with the cost that some
completions may be delayed in case of resends.

In order to reproduce the problem, add "flush_ep" mode to perftest.