UCT/IB/LOG: Fix SG entries logging
Even if we don't want to dump data from all SG elements (for example, in
case of zero copy which may be inaccessible memory), we still want to
dump the meta-data (address/length/lkey).
This change separates the parameter of how SG entries to log, from how
many SG entries to dump data for.