UCP/WIREUP/TEST: Fix endpoint reconfiguration error with many lanes
When the number of selected lanes exceeds UCP_MAX_LANES, we need to stop
adding more lanes and initialize the endpoint correctly.
Before this fix, if we stop adding lanes when this limit is reached, we
could end up selecting less am_bw/rma_bw lanes when connecting to worker
address, than when connecting to endpoint address via wireup protocol.
This is because when connecting to worker address we can potentially
select any available resource, while when connecting to endpoint address
we select only among the already existing lanes. As a result of this
discrepancy, we get a different endpoint configuration after wireup.
In order to solve this, we try to reuse existing lanes whenever we reach
the lane limit. This ensures that connecting to worker address will have
the same result as connecting to endpoint address later on.